// packages/ui/src/hooks/useResize/enum.ts
export enum DeviceBreakpoints {
    MAX_MOBILE = '(max-width: 719px)',
    MIN_TABLET = '(min-width: 720px)',
    MAX_TABLET = '(max-width: 1023px)',
    MIN_DESKTOP = '(min-width: 1024px)',
    MAX_DESKTOP = '(max-width: 1279px)',
    MIN_WIDESCREEN = '(min-width: 1280px)',
    MAX_WIDESCREEN = '(max-width: 1439px)',
    MIN_FULL_HD = '(min-width: 1440px)',
}